.gitignore Added .gitignore, rename companion app
