--- FootTour Companion App/bin/Debug/netcoreapp3.1/FootTour Companion App.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v3.1",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v3.1": {
-      "FootTour Companion App/1.0.0": {
-        "runtime": {
-          "FootTour Companion App.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "FootTour Companion App/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
