--- FootTour Companion App/bin/Debug/netcoreapp3.1/FootTour Companion App.runtimeconfig.dev.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\tankovits.bence\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\tankovits.bence\\.nuget\\packages",
-      "C:\\Program Files (x86)\\Microsoft\\Xamarin\\NuGet"
-    ]
-  }
-}
