+++ FootTourCA/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="FootTour_Companion_App.App"
+﻿<Application x:Class="FootTourCA.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:FootTour_Companion_App"
+             xmlns:local="clr-namespace:FootTourCA"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
