+++ FootTourCA/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace FootTour_Companion_App
+namespace FootTourCA
 {
     /// <summary>
     /// Interaction logic for App.xaml
