+++ FootTourCA/AssemblyInfo.cs
