+++ FootTourCA/MainWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="FootTour_Companion_App.MainWindow"
+﻿<Window x:Class="FootTourCA.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:FootTour_Companion_App"
+        xmlns:local="clr-namespace:FootTourCA"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
     <Grid>
